+++ lib/rdf/data.ex
@@ -37,9 +37,9 @@ defprotocol RDF.Data do
   def include?(data, statements)
 
   @doc """
-  Returns the list of all statements of a RDF data structure.
+  Checks if a RDF data structure contains statements about the given resource.
   """
-  def statements(data)
+  def describes?(data, subject)
 
   @doc """
   Returns a `RDF.Description` of the given subject.
@@ -56,6 +56,11 @@ defprotocol RDF.Data do
   the same subject from all graphs.
   """
   def descriptions(data)
+
+  @doc """
+  Returns the list of all statements of a RDF data structure.
+  """
+  def statements(data)
 
   @doc """
   Returns the set of all resources which are subject of the statements of a RDF data structure.
@@ -129,7 +134,8 @@ defimpl RDF.Data, for: RDF.Description do
   def include?(description, statements),
     do: RDF.Description.include?(description, statements)
 
-  def statements(description), do: RDF.Description.statements(description)
+  def describes?(description, subject),
+    do: RDF.Description.describes?(description, subject)
 
   def description(%RDF.Description{subject: subject} = description, s) do
     with ^subject <- RDF.Statement.convert_subject(s) do
@@ -140,6 +146,8 @@ defimpl RDF.Data, for: RDF.Description do
   end
 
   def descriptions(description), do: [description]
+
+  def statements(description), do: RDF.Description.statements(description)
 
   def subjects(%RDF.Description{subject: subject}), do: MapSet.new([subject])
   def predicates(description), do: RDF.Description.predicates(description)
@@ -193,12 +201,15 @@ defimpl RDF.Data, for: RDF.Graph do
 
   def include?(graph, statements), do: RDF.Graph.include?(graph, statements)
 
-  def statements(graph), do: RDF.Graph.statements(graph)
+  def describes?(graph, subject),
+    do: RDF.Graph.describes?(graph, subject)
 
   def description(graph, subject),
     do: RDF.Graph.description(graph, subject) || RDF.Description.new(subject)
 
   def descriptions(graph), do: RDF.Graph.descriptions(graph)
+
+  def statements(graph), do: RDF.Graph.statements(graph)
 
   def subjects(graph),   do: RDF.Graph.subjects(graph)
   def predicates(graph), do: RDF.Graph.predicates(graph)
@@ -231,7 +242,8 @@ defimpl RDF.Data, for: RDF.Dataset do
 
   def include?(dataset, statements), do: RDF.Dataset.include?(dataset, statements)
 
-  def statements(dataset), do: RDF.Dataset.statements(dataset)
+  def describes?(dataset, subject),
+    do: RDF.Dataset.who_describes(dataset, subject) != []
 
   def description(dataset, subject) do
     with subject = RDF.Statement.convert_subject(subject) do
@@ -249,6 +261,8 @@ defimpl RDF.Data, for: RDF.Dataset do
     |> subjects
     |> Enum.map(&(description(dataset, &1)))
   end
+
+  def statements(dataset), do: RDF.Dataset.statements(dataset)
 
   def subjects(dataset),   do: RDF.Dataset.subjects(dataset)
   def predicates(dataset), do: RDF.Dataset.predicates(dataset)